requirements.txt
python-telegram-bot>=21.0
ydb[yc]
python-dotenv
PyJWT
requests
pytz
fastapi
uvicorn[standard]
yandex-cloud-ml-sdk>=0.17.0
langgraph>=0.2.0
pydantic>=2.0.0
streamlit>=1.28.0
aiohttp>=3.9.0
